レス書き込みに使う処理をクラスに分離
